requirements.txt
requests>=2.31
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
scipy>=1.11
matplotlib>=3.7
seaborn>=0.12
statsmodels>=0.14
tqdm>=4.65
gprofiler-official>=1.0
matplotlib-venn>=0.11
